+++ app/tutorial/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import OnboardingTour from "@/components/tutorial/OnboardingTour";
+
+export default function TutorialPage() {
+  return <OnboardingTour />;
+}